Refactor: Move AudioPage content into hashing feature folder
