+++ lib/features/hashing/logic/AudioManager.dart
