+++ lib/features/hashing/logic/HashAudioManager.dart
